# Allow ISO 8601 formatted string for `fast_string_to_time`