Currently `ISO_DATETIME` regexp parser doesn't allow ISO 8601 formatted
string, it fallbacks to slower `Date._parse` in `fallback_string_to_time`.

This makes `ISO_DATETIME` allows ISO 8601 formatted string for
`fast_string_to_time`, it makes ISO 8601 formatted string parsing about
3.5x faster.

```ruby
type = ActiveRecord::Type.lookup(:datetime)
local = Time.now.iso8601
utc = Time.now.utc.iso8601

Benchmark.ips do |x|
  x.report("type.cast(local)") { type.cast(local); type.cast(local) }
  x.report("type.cast(utc)")   { type.cast(utc);   type.cast(utc)   }
end
```

Before:

```
Warming up --------------------------------------
    type.cast(local)     2.198k i/100ms
      type.cast(utc)     2.443k i/100ms
Calculating -------------------------------------
    type.cast(local)     22.371k (± 5.9%) i/s -    112.098k in   5.028419s
      type.cast(utc)     23.359k (± 5.5%) i/s -    117.264k in   5.034913s
```

After:

```
Warming up --------------------------------------
    type.cast(local)     6.918k i/100ms
      type.cast(utc)     9.414k i/100ms
Calculating -------------------------------------
    type.cast(local)     71.468k (± 6.2%) i/s -    359.736k in   5.053612s
      type.cast(utc)     86.258k (± 4.5%) i/s -    433.044k in   5.029979s
```